# server/api.yml
openapi: "3.0.3"
info:
  title: TechMentorX API
  description: Backend API for TechMentorX student performance tracking, engagement analysis, and mental health support platform.
  version: "1.0.0"
servers:
  - url: http://localhost:5000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /dashboard:
    get:
      summary: Student personal dashboard
      tags: [Student Dashboard]
      security:
        - BearerAuth: []
      description: "Role: student only"
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                type: object
                properties:
                  student:
                    type: object
                    properties:
                      name:
                        type: string
                      email:
                        type: string
                      department:
                        type: string
                      semester:
                        type: integer
                      enrollmentId:
                        type: string
                  performance:
                    type: object
                    properties:
                      overallAverage:
                        type: number
                      totalQuizzesTaken:
                        type: integer
                      subjectPerformance:
                        type: array
                        items:
                          type: object
                          properties:
                            subjectId:
                              type: string
                            subjectName:
                              type: string
                            averageScore:
                              type: number
                            quizzesTaken:
                              type: integer
                  pendingStudyTopics:
                    type: array
                    items:
                      type: object
...
  /teacher/dashboard:
    get:
      summary: Teacher dashboard overview
      tags: [Teacher]
      security:
        - BearerAuth: []
      description: "Role: teacher only"
      responses:
        "200":
          description: Teacher dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  subjects:
                    type: array
                    items:
                      type: object
                  recentLectures:
                    type: array
                    items:
                      type: object
                  totalSubjects:
                    type: integer
                  totalStudents:
                    type: integer
  /teacher/subjects:
    post:
      summary: Create a new subject
      tags: [Teacher]
      security:
        - BearerAuth: []
      description: "Role: teacher only"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, code, department, semester]
              properties:
                name:
                  type: string
                code:
                  type: string
                department:
                  type: string
                semester:
                  type: integer
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  subject:
                    type: object
  /teacher/subjects/{subjectId}/students:
    post:
      summary: Enroll students in a subject
      tags: [Teacher]
      security:
        - BearerAuth: []
      description: "Role: teacher only"
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [studentIds]
              properties:
                studentIds:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Students enrolled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    get:
      summary: List enrolled students for a subject
      tags: [Teacher]
      security:
        - BearerAuth: []
      description: "Role: teacher only"
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student list
          content:
            application/json:
              schema:
                type: object
                properties:
                  students:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        email:
                          type: string
                        enrollmentId:
                          type: string
                        semester:
                          type: integer
                        department:
                          type: string
  /teacher/lectures:
    post:
      summary: Create a lecture (auto-generates quiz if PPT provided)
      tags: [Teacher]
      security:
        - BearerAuth: []
      description: "Role: teacher only. If pptUrl is provided, a 20-minute quiz is auto-generated via AI."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subjectId, title, topicsCovered]
              properties:
                subjectId:
                  type: string
                title:
                  type: string
                description:
                  type: string
                pptUrl:
                  type: string
                  description: URL/path to PPT — triggers quiz generation
                topicsCovered:
                  type: array
                  items:
                    type: string
                date:
                  type: string
                  format: date-time
                duration:
                  type: integer
                  description: Minutes (default 60)
      responses:
        "201":
          description: Lecture created (with optional quiz)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  lecture:
                    type: object
                  quiz:
                    type: object
                    nullable: true
  /teacher/subjects/{subjectId}/engagement:
    get:
      summary: View engagement analytics for a subject
      tags: [Teacher]
      security:
        - BearerAuth: []
      description: "Role: teacher only. Shows topics needing repetition based on OpenCV engagement analysis."
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Engagement analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  subject:
                    type: object
                    properties:
                      name:
                        type: string
                      code:
                        type: string
                  engagementRecords:
                    type: array
                    items:
                      type: object
                  topicSummary:
                    type: array
                    items:
                      type: object
                      properties:
                        topic:
                          type: string
                        avgEngagement:
                          type: number
                        needsRepeat:
                          type: boolean
                  topicsNeedingRepeat:
                    type: array
                    items:
                      type: object
...